Refactored CRUD and graph state handlers
